image table and api refactor
